refactor(test): Transactional 제거 및 DB 클리닝 적용
