+++ src/test/java/com/snackgame/server/applegame/domain/game/AppleGamesTest.java
@@ -6,14 +6,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
 
-import javax.persistence.EntityManager;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.transaction.TestTransaction;
@@ -24,11 +21,12 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.snackgame.server.applegame.domain.apple.Apple;
 import com.snackgame.server.applegame.fixture.TestFixture;
+import com.snackgame.server.support.general.DatabaseCleaningDataJpaTest;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@DataJpaTest
 @EnableJpaAuditing
+@DatabaseCleaningDataJpaTest
 class AppleGamesTest {
 
     @Autowired
@@ -36,9 +34,6 @@ class AppleGamesTest {
     @Autowired
     JdbcTemplate jdbcTemplate;
     ObjectMapper objectMapper = new ObjectMapper();
-
-    @Autowired
-    EntityManager em;
 
     @BeforeEach
     void setUp() {
